Desain UI

+++ app/src/main/java/com/shakilla/penjualan/model/ModelPenjualan.kt
@@ -3,6 +3,6 @@ package com.shakilla.penjualan.model
 data class ModelPenjualan(
     var idPenjualan: String? = null,
     var tanggal: Long? = 0,
-    var total: Int? = 0,
-    var keuntungan: Int? = 0
+    var total: Long? = 0,
+    var keuntungan: Long? = 0
 )
